91906(v3)/requirements.txt
setuptools         65.5.0
certifi            2023.5.7
cffi               1.15.1
charset-normalizer 3.1.0
cryptography       41.0.1
customtkinter      5.1.3
geographiclib      2.0
geopy              2.3.0
numpy              1.25.0
Pillow             9.4.0
pip                23.1.2
PyMySQL            1.0.3
pypiwin32          223
pytz               2023.3
requests           2.31.0
timezonefinder     6.2.0